fix: next migration list bug fix.
